Правки меток, разнес файлы будущих экранов по папкам
