--- Recipes/Recipes/AppConstants.swift
@@ -1,14 +1,0 @@
-// AppConstants.swift
-// Copyright © RoadMap. All rights reserved.
-
-import Foundation
-
-enum AppConstants {
-    static let titleText = "Profile"
-    static let titleFont = "Verdana-Bold"
-    static let titleFontSize = 28
-    static let alertTitle = "Change your name and surname"
-    static let alertTextFieldPlaceholder = "Name Surname"
-    static let alertOkText = "Ok"
-    static let alertCancelText = "Cancel"
-}
