+++ Recipes/Recipes/Favorites Screen/FavoritesViewController.swift
@@ -1,0 +1,12 @@
+// FavoritesViewController.swift
+// Copyright © RoadMap. All rights reserved.
+
+import UIKit
+
+/// Экран избранных рецептов
+final class FavoritesViewController: UIViewController {
+    
+    // MARK: - Public Properties
+    var presenter: FavoritesPresenter?
+    
+}
