+++ Recipes/Recipes/Recipe Screen/RecipeViewController.swift
@@ -1,0 +1,13 @@
+// RecipeViewController.swift
+// Copyright © RoadMap. All rights reserved.
+
+import UIKit
+
+/// Экран рецептов
+final class RecipeViewController: UIViewController {
+    
+    // MARK: - Public Properties
+    var presenter: RecipePresenter?
+
+
+}
